requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
grpcio
protobuf
grpcio-tools
grpcio-health-checking
httpx
stripe
typer
rich
web3
celery
redis
transformers
torch
accelerate
bitsandbytes
